Use a single table for the available optical indexes
Two parts of the user interface kept a separate copy of the bands that each
index needs. The copies were different, so SEPAL did not offer some indexes
that it can calculate. The code gave no error message.
