+++ modules/gui/src/app/home/body/process/panels/inputImageryWithDerived/inputImage.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
+import {getAvailableIndexes} from '~/app/home/body/process/opticalIndexes'
 import {RecipeFormPanel, recipeFormPanel} from '~/app/home/body/process/recipeFormPanel'
 import {compose} from '~/compose'
 import {selectFrom} from '~/stateUtils'
@@ -13,7 +14,6 @@ import {AssetSection} from './assetSection'
 import {BandSetSpec} from './bandSetSpec'
 import {ImageForm} from './imageForm'
 import styles from './inputImage.module.css'
-import {getAvailableIndexes} from './opticalIndexes'
 import {getProfileBandSetSpecs, isProfileDisabled} from './profiles'
 import {RecipeSection} from './recipeSection'
 import {SectionSelection} from './sectionSelection'
